Grouped parameters now supported
1. In the commonly seen case where multiple grids of the same shape and
dtype might be needed (v1, v2 and v3 in Gradient), do not generate
multiple versions of the variables describing the shape in C. The reuse
of the same dimension variables will enable pointer assignments to work.
2. Now that we are maintaining matrix params (function parameters which
are grids) and value params (function parameters that are simple PODs)
separately in FunctionDescriptor, the try catch blocks in
FunctionManager can be removed.
3. Also fixed some lines where the dtype was still hardcoded for the
generated code in FunctionManager.