[WIP] FunctionDescriptor creation is now internal to Propagator
The interface with AcousticWave2D was getting messy due to the
dependence on both Propagator and FunctionDescriptor. Moved the
creation of the FunctionDescriptor inside the Propagator class so
AcousticWave2D does not need to know about FunctionDescriptors any
more.
Propagator now knows about each of the variables inside the sympy
expression it will receive.
This was needed to make the variable substitution for time easier in
the case where we are cycling between three time steps only (not saving
entire time history).