Use a random salt in the filename
This is a temporary fix for the problem that when this process is
launched with mpi, the various workers try to write to the same file
and fail. A better fix would be to use the MPI rank in the filename.
However, there doesn’t seem to be a standard env var where all MPI
distributions provide the rank.