SymbolicData: Add symbolic stencil derivative expressions as properties
This enables the user to define SymPy equations in terms of properties
of the `TimeData` variables, eg. `u.dx2` where `u = TimeData(...)`.
Currently only consists of `u.dx2`, u.dy2`, `u.dt` and `u.forward`,
but this list should be extended soon as we integrate more equations
with this new wrokflow.