SymbolicData: Duplicate type objects in __new__ and invoke __init__
Since `cls` is a class variable we need to duplicate (deep copy) it when
creating new Function objects to enable symbol caching of multiple
data objects of the same type (eg. `DenseData`). For some reason this
also requires us to invoke `__init__` manually on the new Function object.

Importantly it also means that we need to inherit in class hierarchies
via `super(self.__class__, self)` to honour the dynamic sub-typing.