Operator: Shift time-stepping indices in new-style stencil definitions
The Propagaor assumes LHS to in terms of `f[t]` and RHS to be in terms
of `f[t-1]`, `f[t-2]`, etc. However, symmetric time derivatives insert
`f(t+1)` and `f(t+2)` terms, so we need to shift time indices by the
largest `s` coefficient used in the LHS of each equation.