Compiler: Adding new compiler infrastructure that replaces `JitManager`
We now drive JIT compilation with compiler objects that inherit from
`codepy.toolchain.GCCToolchain` and encode our own "known best" compiler
flags for architectures we support. This initial commit only implements
the `GNUCompiler` toolchain, but other will follow shortly.

We also now allow explicit overrides in the `Operator` and `Propagator`
constructors and try to derive a compiler from env vars if these are
not specified. Currently the key variable for this is DEVITO_ARCH.