FWI: Switching core operator definitions to new style
This now bypasses `_init_taylor` and `smart_sympy_replace` for the
core forward and backward wave equation stencils. The new stencils
are defined purely from symbolic terms provided by the symbolic data
objects, while the insertion and reset steps for Gradient and Born
are still defined in the Propagator's "indexed" notation.

Note that the source and receiver insertion now happens after the
initial Operator setup, so that the Operator doesn;t need to pass
things on explicitly.

Also note that auto-detection of the kernel signature (`input_params`
and `output_params`) is still flawed, so we need to be explicit for
all except the Forward operator.