Iteration: Add new Propagator-like class for individual loop iterations
An `Iteration` encapsulates a single loop around a set of stencil
expressions. This allows users to manually insert loops with
sympy expressions through the lower-level Propagator API. Ultimately,
this could also be used by the Propagator itself to compose the inner
loop nest. It is also the first step towards dealing with symbolic
dimensions and loops more elegantly.