StencilKernel: New code generator based on Iterations and Expressions
The new `StencilKernel` class is intended to work similar to the current
Propagator, but based on building generic loop nests according to the
dimensions and function definition in the given stencil expressions. This
new code generator is based on composing `Iteration` and `Expression`
objects and delegates code generation and symbol substitution accordingly.