StencilKernel: Utilise visitors to resolve and substitute loop variables
Instead of mangling the index variable in the `Iteration` constructor
we now derive loop variables from loop dimensions in a visitor that
also inserts the explicit buffer indices when we encounter buffered
dimensions. This visitor also builds the required symmbol substitutions,
which are duly applied by the new `ExpressionSubstitute` visitor.