dle+StencilKernel: Fix argument binding
Previously, we were hitting a cache-related bug, due to instantiating
a Dimension object passing in the same name used for another Dimension
(the blocked one). This turned out to lead to cache hits and the loss
of information (eg, the size of a Dimension, as we were passing no size
in the newly created Dimension).