TestRewriter: Tweaks to handle new operator structure and xfail one test
The temporaries test uses `operator.stencils`, which is not stored by
the new-style operators anymore, so we xfail for now.

The other tests relied on the implicit return of symbolic data values,
which is also not supported anymore. Small tweaks are required to get
the data flow right again. This will be cleaned up further once the
seismic example operators are restructured.